issue 2689 reorganize api classes
